--- apps/lib/constants/app_constants.dart
@@ -1,1 +1,0 @@
-const appName = "Wellness";
