+++ apps/lib/constants/colors.dart
@@ -1,0 +1,3 @@
+import 'package:flutter/material.dart';
+
+const primaryColor = const Color(0xFF1F96F3);
