+++ apps/lib/constants/versioning.dart
@@ -1,0 +1,1 @@
+const appName = "Wellnes";
