+++ apps/lib/main.dart
@@ -1,4 +1,5 @@
-import 'package:apps/pages/home_page.dart';
+import 'package:apps/constants/versioning.dart';
+import 'package:apps/pages/home/home_page.dart';
 import 'package:flutter/material.dart';
 
 void main() {
@@ -9,12 +10,9 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-        visualDensity: VisualDensity.adaptivePlatformDensity,
-      ),
-      home: HomePage(title: 'Flutter Demo Home Page'),
+      title: appName,
+      debugShowCheckedModeBanner: false,
+      home: HomePage(),
     );
   }
 }
