+++ apps/lib/pages/home/home_page.dart
@@ -1,0 +1,67 @@
+import 'package:apps/constants/colors.dart';
+import 'package:flutter/material.dart';
+
+class HomePage extends StatefulWidget {
+  HomePage({Key key}) : super(key: key);
+
+  @override
+  _HomePageState createState() => _HomePageState();
+}
+
+class _HomePageState extends State<HomePage> {
+  int _counter = 0;
+
+  void _incrementCounter() {
+    setState(() {
+      _counter++;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: _buildAppBar(),
+      body: Center(
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            Text(
+              'You have pushed the button this many times:',
+            ),
+            Text(
+              '$_counter',
+              style: Theme.of(context).textTheme.headline4,
+            ),
+          ],
+        ),
+      ),
+      bottomNavigationBar: _buildBottomNavigationBar(),
+    );
+  }
+
+  AppBar _buildAppBar() {
+    return AppBar(
+      title: Text(
+        "Wellness Center",
+      ),
+      backgroundColor: primaryColor,
+      actions: [
+        IconButton(icon: Icon(Icons.notifications), onPressed: () {}),
+        IconButton(icon: Icon(Icons.message), onPressed: () {}),
+        Padding(
+          padding: const EdgeInsets.all(8.0),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              CircleAvatar(
+                radius: 15,
+                backgroundImage: NetworkImage(
+                    "https://www.inpixio.com/remove-background/images/animation-light-square.gif"),
+              )
+            ],
+          ),
+        )
+      ],
+    );
+  }
+}
